+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to the "vscode-camelk" extension will be documented in this 
 ## 0.0.26
 
 - Adapt Didact tutorial registration to VS Code Didact 0.4.0 updated API
+- Update default runtime version to v1.4.1
 
 ## 0.0.25