+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to the "vscode-camelk" extension will be documented in this 
 ## 0.0.26
 
 - Adapt Didact tutorial registration to VS Code Didact 0.4.0 updated API
-- Update default runtime version to v1.4.1
+- Update default runtime version to v1.5.0
 - `resource` has been replaced by `resources` to allow providing several of them in Camel K VS Code Tasks
 
 ## 0.0.25